+++ Less3/Classes/AuthManager.cs
@@ -887,7 +887,7 @@ namespace Less3.Classes
 
             return md;
         }
-         
+          
         #endregion
 
         #region Private-Methods