+++ Less3/Common.cs
@@ -981,23 +981,8 @@ namespace Less3
 
         public static T DeserializeJson<T>(string json)
         {
-            if (String.IsNullOrEmpty(json)) throw new ArgumentNullException(nameof(json));
-
-            try
-            {
-                return JsonConvert.DeserializeObject<T>(json);
-            }
-            catch (Exception e)
-            {
-                Console.WriteLine("");
-                Console.WriteLine("Exception while deserializing:");
-                Console.WriteLine(json);
-                Console.WriteLine("");
-                Console.WriteLine("Exception:");
-                Console.WriteLine(SerializeJson(e, true));
-                Console.WriteLine("");
-                throw e;
-            }
+            if (String.IsNullOrEmpty(json)) throw new ArgumentNullException(nameof(json)); 
+            return JsonConvert.DeserializeObject<T>(json); 
         }
 
         public static T DeserializeJson<T>(byte[] data)
@@ -1099,44 +1084,27 @@ namespace Less3
 
         #region Crypto
 
-        public static string Md5(byte[] data)
+        public static byte[] Md5(byte[] data)
         {
             if (data == null) return null;
-
-            MD5 md5 = MD5.Create();
-            byte[] hash = md5.ComputeHash(data);
-            StringBuilder sb = new StringBuilder();
-            for (int i = 0; i < hash.Length; i++) sb.Append(hash[i].ToString("X2"));
-            string ret = sb.ToString();
-            return ret;
+            return MD5.Create().ComputeHash(data);
         }
 
-        public static string Md5(string data)
+        public static byte[] Md5(string data)
         {
             if (String.IsNullOrEmpty(data)) return null;
-
-            MD5 md5 = MD5.Create();
-            byte[] dataBytes = System.Text.Encoding.ASCII.GetBytes(data);
-            byte[] hash = md5.ComputeHash(dataBytes);
-            StringBuilder sb = new StringBuilder();
-            for (int i = 0; i < hash.Length; i++) sb.Append(hash[i].ToString("X2"));
-            string ret = sb.ToString();
-            return ret;
+            return Md5(Encoding.UTF8.GetBytes(data));
         }
 
-        public static string Md5(Stream stream)
+        public static byte[] Md5(Stream stream)
         {
             if (stream == null || !stream.CanRead) return null;
 
             MD5 md5 = MD5.Create();
-            byte[] hash = md5.ComputeHash(stream);
-            StringBuilder sb = new StringBuilder();
-            for (int i = 0; i < hash.Length; i++) sb.Append(hash[i].ToString("X2"));
-            string ret = sb.ToString();
-            return ret;
+            return md5.ComputeHash(stream);
         }
 
-        public static async Task<string> Md5Async(Stream stream, int bufferSize)
+        public static async Task<byte[]> Md5Async(Stream stream, int bufferSize)
         {
             using (var md5 = MD5.Create())
             {
@@ -1153,8 +1121,34 @@ namespace Less3
                 } while (read > 0);
 
                 md5.TransformFinalBlock(buffer, 0, 0);
-                return BitConverter.ToString(md5.Hash).Replace("-", "").ToUpperInvariant();
+                return md5.Hash; 
             }
+        }
+
+        public static byte[] HmacSha1(byte[] input, byte[] key)
+        {
+            if (input == null) return null;
+            if (key == null || key.Length < 1) return null;
+            return new HMACSHA1(key).ComputeHash(input);
+        }
+
+        public static byte[] HmacSha256(byte[] input, byte[] key)
+        {
+            if (input == null) return null;
+            if (key == null || key.Length < 1) return null;
+            return new HMACSHA256(key).ComputeHash(input);
+        }
+
+        public static byte[] Sha1(byte[] input)
+        {
+            if (input == null) return null;
+            return new SHA1Managed().ComputeHash(input);
+        }
+
+        public static byte[] Sha256(byte[] data)
+        {
+            if (data == null) return null;
+            return new SHA256Managed().ComputeHash(data);
         }
 
         #endregion